+++ generator/src/main/java/com/github/yituhealthcare/arc/generator/codegen/TypeGenerator.java
@@ -2,10 +2,11 @@ package com.github.yituhealthcare.arc.generator.codegen;
 
 import com.github.yituhealthcare.arc.dgraph.annotation.DgraphType;
 import com.github.yituhealthcare.arc.dgraph.annotation.UidField;
-import com.github.yituhealthcare.arc.dgraph.dictionary.IDgraphType;
+import com.github.yituhealthcare.arc.dgraph.dictionary.IDomainClass;
 import com.github.yituhealthcare.arc.generator.codegen.spec.FieldSpecGenAbstractGetter;
 import com.github.yituhealthcare.arc.generator.codegen.spec.FieldSpecGenGetter;
 import com.github.yituhealthcare.arc.generator.codegen.spec.FieldSpecGenSetter;
+import com.github.yituhealthcare.arc.generator.codegen.spec.TypeSpecGenBuilder;
 import com.github.yituhealthcare.arc.generator.codegen.util.JavadocUtils;
 import com.github.yituhealthcare.arc.generator.codegen.util.PackageManager;
 import com.github.yituhealthcare.arc.generator.codegen.util.SpecNameManager;
@@ -55,33 +56,38 @@ public class TypeGenerator implements IGenerator {
         final Predicate<ObjectTypeDefinition> isOperator = new IsOperator(GraphqlTypeUtils.getOperationTypeNames(typeDefinitionRegistry));
         final GraphqlType2JavapoetTypeName toJavapoetTypeName = new GraphqlType2JavapoetTypeName(packageManager);
         final MethodSpecBuilder toMethodSpec = new MethodSpecBuilder(toJavapoetTypeName);
-        final DgraphTypeFiller dgraphTypeFiller = new DgraphTypeFiller(toJavapoetTypeName);
-        final AutowiredFieldFiller autowiredFieldFiller = new AutowiredFieldFiller(packageManager);
+        final IsGraphqlMethodField isGraphqlMethodField = new IsGraphqlMethodField(typeDefinitionRegistry);
+        final DgraphTypeFiller dgraphTypeFiller = new DgraphTypeFiller(toJavapoetTypeName, isGraphqlMethodField);
+        final IsContainsGraphqlMethodField isContainGraphqlMethodField = new IsContainsGraphqlMethodField(isGraphqlMethodField);
+        final AutowiredFieldFiller autowiredFieldFiller = new AutowiredFieldFiller(isContainGraphqlMethodField, packageManager);
         final TypeSpecConvert typeSpecConvert = new TypeSpecConvert(toJavapoetTypeName);
+        final TypeSpecGenBuilder typeSpecGenBuilder = new TypeSpecGenBuilder();
 
         return typeDefinitionRegistry.types().values().stream()
                 .filter(it -> Arrays.stream(SUPPORT_TYPES).anyMatch(t -> t.isInstance(it)))
                 .map(typeDefinition -> {
                     switch (TypeKind.getTypeKind(typeDefinition)) {
                         case Operation:
-                            return new OperatorGenerator(typeSpecConvert, autowiredFieldFiller, toMethodSpec).apply((ObjectTypeDefinition)typeDefinition);
+                            return new OperatorGenerator(typeSpecConvert, autowiredFieldFiller, toMethodSpec).apply((ObjectTypeDefinition) typeDefinition);
                         case Object:
-                            if (isOperator.test((ObjectTypeDefinition)typeDefinition)) {
-                                return new OperatorGenerator(typeSpecConvert, autowiredFieldFiller, toMethodSpec).apply((ObjectTypeDefinition)typeDefinition);
+                            if (isOperator.test((ObjectTypeDefinition) typeDefinition)) {
+                                return new OperatorGenerator(typeSpecConvert, autowiredFieldFiller, toMethodSpec).apply((ObjectTypeDefinition) typeDefinition);
                             } else {
-                                return new ObjectGenerator(typeSpecConvert, dgraphTypeFiller, autowiredFieldFiller, toMethodSpec).apply((ObjectTypeDefinition)typeDefinition);
+                                TypeSpec.Builder typeSpecBuilder = new ObjectGenerator(typeSpecConvert, dgraphTypeFiller, autowiredFieldFiller, toMethodSpec, isGraphqlMethodField).apply((ObjectTypeDefinition) typeDefinition);
+                                return typeSpecGenBuilder.apply(typeSpecBuilder.build()).toBuilder();
                             }
                         case Interface:
                             return new InterfaceGenerator(toJavapoetTypeName).apply((InterfaceTypeDefinition) typeDefinition);
                         case Union:
-                            return new UnionGenerator().apply((UnionTypeDefinition)typeDefinition);
+                            return new UnionGenerator().apply((UnionTypeDefinition) typeDefinition);
                         case Enum:
                         case Scalar:
                         case InputObject:
                         default:
                             throw new IllegalArgumentException("type generator not support this graphql type!");
                     }
-                }).map(it -> JavaFile.builder(packageManager.getTypePackage(), it.build()).build());
+                })
+                .map(it -> JavaFile.builder(packageManager.getTypePackage(), it.build()).build());
     }
 
     static class ObjectGenerator implements Function<ObjectTypeDefinition, TypeSpec.Builder> {
@@ -90,20 +96,22 @@ public class TypeGenerator implements IGenerator {
         private final DgraphTypeFiller dgraphTypeFiller;
         private final AutowiredFieldFiller autowiredFieldFiller;
         private final MethodSpecBuilder toMethodSpec;
+        private final IsGraphqlMethodField isGraphqlMethodField;
 
         public ObjectGenerator(TypeSpecConvert typeSpecConvert,
                                DgraphTypeFiller dgraphTypeFiller,
                                AutowiredFieldFiller autowiredFieldFiller,
-                               MethodSpecBuilder toMethodSpec) {
+                               MethodSpecBuilder toMethodSpec, IsGraphqlMethodField isGraphqlMethodField) {
             this.typeSpecConvert = typeSpecConvert;
             this.dgraphTypeFiller = dgraphTypeFiller;
             this.autowiredFieldFiller = autowiredFieldFiller;
             this.toMethodSpec = toMethodSpec;
+            this.isGraphqlMethodField = isGraphqlMethodField;
         }
 
         @Override
         public TypeSpec.Builder apply(ObjectTypeDefinition objectTypeDefinition) {
-            return new MethodsFiller(toMethodSpec, new IsGraphqlMethodField())
+            return new MethodsFiller(toMethodSpec, isGraphqlMethodField)
                     .andThen(dgraphTypeFiller)
                     .andThen(autowiredFieldFiller)
                     .apply(typeSpecConvert.apply(objectTypeDefinition), objectTypeDefinition);
@@ -183,10 +191,8 @@ public class TypeGenerator implements IGenerator {
         public MethodSpec apply(ObjectTypeDefinition objectTypeDefinition, FieldDefinition fieldDefinition) {
             return MethodSpec.methodBuilder(fieldDefinition.getName())
                     .addModifiers(Modifier.PUBLIC)
-                    .beginControlFlow("return dataFetchingEnvironment -> ")
                     .addAnnotation(AnnotationSpec.builder(GraphqlMethod.class).addMember("type", "$S", objectTypeDefinition.getName()).build())
-                    .addStatement(" return $L.handle$L(dataFetchingEnvironment)", SpecNameManager.getUnCapitalizeApiName(objectTypeDefinition), StringUtils.capitalize(fieldDefinition.getName()))
-                    .addStatement("}")
+                    .addStatement("return dataFetchingEnvironment -> $L.handle$L(dataFetchingEnvironment)", SpecNameManager.getUnCapitalizeApiName(objectTypeDefinition), StringUtils.capitalize(fieldDefinition.getName()))
                     .returns(ParameterizedTypeName.get(ClassName.get(DataFetcher.class), toJavapoetTypeName.apply(fieldDefinition.getType())))
                     .build();
         }
@@ -204,7 +210,7 @@ public class TypeGenerator implements IGenerator {
             return TypeSpec.classBuilder(StringUtils.capitalize(objectTypeDefinition.getName()))
                     .addModifiers(Modifier.PUBLIC)
                     .addAnnotation(Graphql.class)
-                    .addSuperinterface(IDgraphType.class)
+                    .addSuperinterface(IDomainClass.class)
                     .addSuperinterfaces(objectTypeDefinition.getImplements().stream()
                             .map(it -> toJavapoetTypeName.apply(GraphqlTypeUtils.getUnWrapperType(it)))
                             .collect(Collectors.toList()))
@@ -215,12 +221,13 @@ public class TypeGenerator implements IGenerator {
     static class DgraphTypeFiller implements Filler<TypeSpec.Builder, ObjectTypeDefinition> {
 
         private final GraphqlType2JavapoetTypeName toJavapoetTypeName;
-        private final IsGraphqlMethodField isGraphqlMethodField = new IsGraphqlMethodField();
+        private final IsGraphqlMethodField isGraphqlMethodField;
         private final FieldSpecGenGetter fieldSpecGenGetter = new FieldSpecGenGetter();
         private final FieldSpecGenSetter fieldSpecGenSetter = new FieldSpecGenSetter();
 
-        public DgraphTypeFiller(GraphqlType2JavapoetTypeName toJavapoetTypeName) {
+        public DgraphTypeFiller(GraphqlType2JavapoetTypeName toJavapoetTypeName, IsGraphqlMethodField isGraphqlMethodField) {
             this.toJavapoetTypeName = toJavapoetTypeName;
+            this.isGraphqlMethodField = isGraphqlMethodField;
         }
 
         @Override
@@ -251,10 +258,11 @@ public class TypeGenerator implements IGenerator {
 
     static class AutowiredFieldFiller implements Filler<TypeSpec.Builder, ObjectTypeDefinition> {
 
-        private final Predicate<ObjectTypeDefinition> isContainGraphqlMethodField = new IsContainsGraphqlMethodField();
+        private final Predicate<ObjectTypeDefinition> isContainGraphqlMethodField;
         private final PackageManager packageManager;
 
-        public AutowiredFieldFiller(PackageManager packageManager) {
+        public AutowiredFieldFiller(Predicate<ObjectTypeDefinition> isContainGraphqlMethodField, PackageManager packageManager) {
+            this.isContainGraphqlMethodField = isContainGraphqlMethodField;
             this.packageManager = packageManager;
         }