+++ generator/src/main/java/com/github/yituhealthcare/arc/generator/dictionary/GeneratorGlobalConst.java
@@ -2,6 +2,10 @@ package com.github.yituhealthcare.arc.generator.dictionary;
 
 import com.squareup.javapoet.ClassName;
 import com.squareup.javapoet.CodeBlock;
+import com.squareup.javapoet.ParameterizedTypeName;
+import com.squareup.javapoet.TypeName;
+
+import java.util.Map;
 
 /**
  * 代码生成器全局配置
@@ -11,11 +15,18 @@ import com.squareup.javapoet.CodeBlock;
  */
 public class GeneratorGlobalConst {
 
-    public static final CodeBlock GENERAL_CODE_BLOCK = CodeBlock.builder().add("Generate with GraphQL Schema By Arc").build();
+    public static final CodeBlock GENERAL_CODE_BLOCK = CodeBlock.builder()
+            .add("Generate with GraphQL Schema")
+            .add("\r\n")
+            .add("\r\n")
+            .add("@author Arc")
+            .build();
 
     public static final ClassName DEFAULT_LIST_CLASS_NAME = ClassName.get("java.util", "List");
 
-    public static final String INTERFACE_SUFFIX = "Service";
+    public static final TypeName DEFAULT_MAP_TYPE = ParameterizedTypeName.get(Map.class, String.class, Object.class);
+
+    public static final String INTERFACE_SUFFIX = "DataFetcher";
 
     public static final String REPOSITORY_SUFFIX = "Repository";