+++ generator/src/test/resources/api_test_QueryService.java
@@ -1,9 +1,0 @@
-package a.b.c.api;
-import a.b.c.type.Project;
-import graphql.schema.DataFetchingEnvironment;
-/**
- * Generate with GraphQL Schema By Arc
- */
-public interface QueryService {
-  Project handleProject(DataFetchingEnvironment dataFetchingEnvironment);
-}