+++ generator/src/test/resources/dictionary_test_E1.java
@@ -1,7 +1,9 @@
 package a.b.c.dictionary;
 /**
  * this is e1
- * Generate with GraphQL Schema By Arc
+ * Generate with GraphQL Schema
+ *
+ * @author Arc
  */
 public enum E1 {
   /**