+++ generator/src/test/resources/type_test_Query.java
@@ -1,7 +1,7 @@
 package a.b.c.type;
 
-import a.b.c.api.QueryService;
-import com.github.yituhealthcare.arc.dgraph.dictionary.IDgraphType;
+import a.b.c.datafetcher.QueryDataFetcher;
+import com.github.yituhealthcare.arc.dgraph.dictionary.IDomainClass;
 import com.github.yituhealthcare.arc.graphql.annotation.Graphql;
 import com.github.yituhealthcare.arc.graphql.annotation.GraphqlMethod;
 import graphql.schema.DataFetcher;
@@ -10,29 +10,27 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 /**
  * Query
- * Generate with GraphQL Schema By Arc
+ * Generate with GraphQL Schema
+ *
+ * @author Arc
  */
 @Graphql
-public class Query implements IDgraphType {
+public class Query implements IDomainClass {
   @Autowired
-  private QueryService queryService;
+  private QueryDataFetcher queryDataFetcher;
 
   @GraphqlMethod(
       type = "Query"
   )
   public DataFetcher<Project> project() {
-    return dataFetchingEnvironment ->  {
-      return queryService.handleProject(dataFetchingEnvironment);
-    };
+    return dataFetchingEnvironment -> queryDataFetcher.handleProject(dataFetchingEnvironment);
   }
 
   @GraphqlMethod(
       type = "Query"
   )
   public DataFetcher<List<User>> users() {
-    return dataFetchingEnvironment ->  {
-      return queryService.handleUsers(dataFetchingEnvironment);
-    };
+    return dataFetchingEnvironment -> queryDataFetcher.handleUsers(dataFetchingEnvironment);
   }
 
 }