+++ generator/src/test/resources/type_test_User.java
@@ -1,21 +1,30 @@
 package a.b.c.type;
 
 import com.github.yituhealthcare.arc.dgraph.annotation.DgraphType;
-import com.github.yituhealthcare.arc.dgraph.dictionary.IDgraphType;
+import com.github.yituhealthcare.arc.dgraph.dictionary.IDomainClass;
 import com.github.yituhealthcare.arc.graphql.annotation.Graphql;
 import java.lang.String;
 
 /**
  * User
- * Generate with GraphQL Schema By Arc
+ * Generate with GraphQL Schema
+ *
+ * @author Arc
  */
 @Graphql
 @DgraphType("USER")
-public class User implements IDgraphType {
+public class User implements IDomainClass {
   /**
    * name
    */
   private String name;
+
+  public User(String name) {
+    this.name = name;
+  }
+
+  public User() {
+  }
 
   public void setName(String name) {
     this.name = name;
@@ -25,4 +34,23 @@ public class User implements IDgraphType {
     return name;
   }
 
+  public static UserBuilder builder() {
+    return new UserBuilder();
+  }
+
+  public static class UserBuilder {
+    private String name;
+
+    private UserBuilder() {
+    }
+
+    public UserBuilder name(String name) {
+      this.name = name;
+      return this;
+    }
+
+    public User build() {
+      return new User(name);
+    }
+  }
 }