+++ generator/src/test/resources/type_union_test_Entity.java
@@ -2,7 +2,9 @@ package a.b.c.type;
 
 /**
  * Entity
- * Generate with GraphQL Schema By Arc
+ * Generate with GraphQL Schema
+ *
+ * @author Arc
  */
 public abstract class Entity {
 }