+++ generator/src/test/resources/type_union_test_Project.java
@@ -2,17 +2,19 @@ package a.b.c.type;
 
 import com.github.yituhealthcare.arc.dgraph.annotation.DgraphType;
 import com.github.yituhealthcare.arc.dgraph.annotation.UidField;
-import com.github.yituhealthcare.arc.dgraph.dictionary.IDgraphType;
+import com.github.yituhealthcare.arc.dgraph.dictionary.IDomainClass;
 import com.github.yituhealthcare.arc.graphql.annotation.Graphql;
 import java.lang.String;
 
 /**
  * Project
- * Generate with GraphQL Schema By Arc
+ * Generate with GraphQL Schema
+ *
+ * @author Arc
  */
 @Graphql
 @DgraphType("PROJECT")
-public class Project implements IDgraphType {
+public class Project implements IDomainClass {
   /**
    * id
    */
@@ -23,6 +25,14 @@ public class Project implements IDgraphType {
    * name
    */
   private String name;
+
+  public Project(String id, String name) {
+    this.id = id;
+    this.name = name;
+  }
+
+  public Project() {
+  }
 
   public void setId(String id) {
     this.id = id;
@@ -38,5 +48,32 @@ public class Project implements IDgraphType {
 
   public String getName() {
     return name;
+  }
+
+  public static ProjectBuilder builder() {
+    return new ProjectBuilder();
+  }
+
+  public static class ProjectBuilder {
+    private String id;
+
+    private String name;
+
+    private ProjectBuilder() {
+    }
+
+    public ProjectBuilder id(String id) {
+      this.id = id;
+      return this;
+    }
+
+    public ProjectBuilder name(String name) {
+      this.name = name;
+      return this;
+    }
+
+    public Project build() {
+      return new Project(id,name);
+    }
   }
 }